Done some minor device group refactoring
